+++ src/org/usfirst/frc/team6814/robot/subsystems/DriveFunctions.java
@@ -1,34 +1,31 @@
 package org.usfirst.frc.team6814.robot.subsystems;
 
-import org.usfirst.frc.team6814.robot.RobotMap;
 import org.usfirst.frc.team6814.robot.commands.EncoderBackward;
 import org.usfirst.frc.team6814.robot.commands.EncoderForward;
 import org.usfirst.frc.team6814.robot.commands.EncoderLeft;
 import org.usfirst.frc.team6814.robot.commands.EncoderRight;
 
-import edu.wpi.first.wpilibj.Encoder;
-
 public class DriveFunctions {
-	public static void driveForward(double speed, int feet, int inches) {
+	public static void driveForward(double speed, double feet, double inches) {
 		
 		EncoderForward encForward = new EncoderForward(speed, feet, inches);
 		
 		
 	}
 	
-	public static void driveBackward(double speed, int feet, int inches) {
+	public static void driveBackward(double speed, double feet, double inches) {
 		
 		EncoderBackward encBackward = new EncoderBackward(speed, feet, inches);
 		
 	}
 	
-	public static void driveLeft(double speed, int feet, int inches) {
+	public static void driveLeft(double speed, double feet, double inches) {
 		
 		EncoderLeft encleft = new EncoderLeft(speed, feet, inches);
 		
 	}
 	
-	public static void driveRight(double speed, int feet, int inches) {
+	public static void driveRight(double speed, double feet, double inches) {
 	
 		EncoderRight encRight = new EncoderRight(speed, feet, inches);
